add test cases
