+++ Battleships/AsciiGameUserIntefaceTests/AsciiGameUserInterfaceTestExtensions.cs
@@ -1,0 +1,8 @@
+﻿using Battleships;
+using Moq;
+
+namespace AsciiGameUserIntefaceTests;
+
+public static class AsciiGameUserInterfaceTestExtensions
+{
+}
