+++ Battleships/AsciiGameUserIntefaceTests/GameInterface.cs
@@ -1,0 +1,32 @@
+﻿using NUnit.Framework;
+
+namespace AsciiGameUserIntefaceTests;
+
+[TestFixture]
+public class GameInterface : Spec
+{
+    [Ignore("Not implemented")]
+    [Test]
+    public void Draws_ascii_board() =>
+        Assert.Fail();
+
+    [Ignore("Not implemented")]
+    [Test]
+    public void Draws_missed_squares() =>
+        Assert.Fail();
+
+    [Ignore("Not implemented")]
+    [Test]
+    public void Draws_hit_squares() =>
+        Assert.Fail();
+
+    [Ignore("Not implemented")]
+    [Test]
+    public void Shows_game_status() =>
+        Assert.Fail();
+
+    [Ignore("Not implemented")]
+    [Test]
+    public void Shows_game_message() =>
+        Assert.Fail();
+}
