+++ Battleships/AsciiGameUserInterface/AsciiMiss.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Battleships;
+
+public class AsciiMiss : IAsciiGameObject
+{
+    public char[,] Graphics => throw new NotImplementedException();
+}
