+++ Battleships/AsciiGameUserInterface/IAsciiGameObject.cs
@@ -1,0 +1,6 @@
+﻿namespace Battleships;
+
+public interface IAsciiGameObject : IGameObject
+{
+    char[,] Graphics { get; }
+}
