+++ Battleships/Battleships/IGameUserInterface.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Collections.Generic;
 
 namespace Battleships;
 
@@ -42,129 +40,4 @@ public interface IBattleShipInputController : IGameInputController
     public Game? Game { get; }
     void AddGame(Game game);
     void RegisterFireMissileEvent(Coordinate coordinate);
-}
-
-public class ConsoleGameInputController : IBattleShipInputController
-{
-    public Game Game { get; private set; } = default!;
-
-    public void AddGame(Game game)
-    {
-        Game = game;
-    }
-
-    public void RegisterFireMissileEvent(Coordinate coordinate) =>
-        Game.OnFireMissile(coordinate);
-}
-
-public class AsciiGameUserInterface<TGameObjectFactory> : IGameUserInterface<TGameObjectFactory> where TGameObjectFactory : IGameObjectFactory
-{
-    private readonly IConsole console;
-    private readonly char[,] screen;
-
-    private string message;
-
-    public List<StatusField> Status => new();
-
-    public TGameObjectFactory GameObjectFactory { get; }
-    public IGameInputController GameInputController { get; }
-
-    public AsciiGameUserInterface(IConsole console, int width, int height, TGameObjectFactory gameObjectFactory, IBattleShipInputController gameInputController)
-    {
-        this.console = console;
-        this.screen = new char[width, height];
-
-        GameObjectFactory = gameObjectFactory;
-        GameInputController = gameInputController;
-    }
-
-    public void Draw(IGameObject gameObject, Coordinate coordinate)
-    {
-        if (gameObject is not IAsciiGameObject asciiGameObject) { return; }
-
-        var graphics = asciiGameObject.Graphics;
-
-        for (int x = 0; x < graphics.GetLength(0); x++)
-        {
-            for (int y = 0; y < graphics.GetLength(1); y++)
-            {
-                screen[x + coordinate.XPos, y + coordinate.YPos] = graphics[x, y];
-            }
-        }
-    }
-
-    public void ShowMessage(string text) =>
-        message = text;
-
-    public void Paint()
-    {
-        console.Clear();
-
-        console.WriteLine(message);
-
-        for (int x = 0; x < screen.GetLength(0); x++)
-        {
-            for (int y = 0; y < screen.GetLength(1); y++)
-            {
-                if (screen[x, y] is default(char)) { continue; }
-
-                console.Write(screen[x, y]);
-                screen[x, y] = default;
-            }
-
-            console.WriteLine("");
-        }
-
-        console.WriteLine(string.Join(" | ", Status.Select(s => $"{s.Label} : {s.Value}")));
-    }
-}
-
-public interface IAsciiGameObject : IGameObject
-{
-    char[,] Graphics { get; }
-}
-
-public class AsciiBattleShipGameObjectFactory : IBattleshipGameObjectFactory
-{
-    public IGameObject CreateBoard(int width, int height) =>
-        new AsciiBoard(width, height);
-
-    public IGameObject CreateShip(bool[] status, Direction direction) =>
-        new AsciiShip(status, direction);
-    public IGameObject CreateMiss() =>
-        new AsciiMiss();
-}
-
-public class AsciiBoard : IAsciiGameObject
-{
-    private int width;
-    private int height;
-
-    public AsciiBoard(int width, int height)
-    {
-        this.width = width;
-        this.height = height;
-    }
-
-    public char[,] Graphics => throw new NotImplementedException();
-}
-
-public class AsciiShip : IAsciiGameObject
-{
-    private Direction direction;
-
-    public AsciiShip(bool[] status, Direction direction)
-    {
-        Status = status;
-        this.direction = direction;
-    }
-
-    public bool[] Status { get; }
-
-    public char[,] Graphics => throw new NotImplementedException();
-}
-
-public class AsciiMiss : IAsciiGameObject
-{
-    public char[,] Graphics => throw new NotImplementedException();
 }
