+++ Battleships/Battleships/Program.cs
@@ -1,14 +1,8 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-
-namespace Battleships;
+﻿namespace Battleships;
 
 internal class Program
 {
-    static async Task Main(string[] args)
+    static void Main(string[] args)
     {
-        GameRunner gameSession = new GameRunner(new GameManagerFactory().Create(new AsciiGameUserInterface<IBattleshipGameObjectFactory>(new SystemConsole(), 10, 10, new AsciiBattleShipGameObjectFactory(), new ConsoleGameInputController())));
-
-        await gameSession.Play();
     }
 }
