+++ Battleships/ConsoleGameInputController/ConsoleGameInputController.csproj
@@ -7,4 +7,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Battleships\Battleships.csproj" />
+  </ItemGroup>
+
 </Project>
