+++ Battleships/ConsoleGameInputControllerTests/ConsoleGameInputControllerTestExtensions.cs
@@ -1,0 +1,8 @@
+﻿using Battleships;
+using Moq;
+
+namespace ConsoleGameInputControllerTests;
+
+public static class ConsoleGameInputControllerTestExtensions
+{
+}
