+++ Battleships/ConsoleGameInputControllerTests/ConsoleInput.cs
@@ -1,0 +1,22 @@
+﻿using NUnit.Framework;
+
+namespace ConsoleGameInputControllerTests;
+
+[TestFixture]
+public class ConsoleInput : Spec
+{
+    [Ignore("Not implemented")]
+    [Test]
+    public void Gets_input_from_console() =>
+        Assert.Fail();
+
+    [Ignore("Not implemented")]
+    [Test]
+    public void Maps_input_to_coordinate() =>
+        Assert.Fail();
+
+    [Ignore("Not implemented")]
+    [Test]
+    public void Registers_a_fire_missile_event() =>
+        Assert.Fail();
+}
